add logger module
