+++ embedded/esp8266/ino/lib/common/errors.h
@@ -1,0 +1,6 @@
+#ifndef LIB_COMMON_ERRORS_H
+#define LIB_COMMON_ERRORS_H
+
+void throw_blocking(const char* caller, const char* message);
+
+#endif // LIB_COMMON_ERRORS_H
