+++ embedded/esp8266/ino/user.example.mk
@@ -3,3 +3,6 @@ ARDUINO_CLI := arduino-cli
 
 # COM port of device
 PORT := /dev/ttyUSB0
+
+# Uncomment to activate serial debugging
+# DEBUG := 1
